% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.find.extr}
\alias{cheb.find.extr}
\title{Curve Fitting with Chebyshev Polynomials and Finding of its Roots}
\usage{
cheb.find.extr(d, x.axis, n, roots.bound.l = NA, roots.bound.u = NA)
}
\arguments{
\item{d}{Zu fittender Datensatz/Zeitreihe (Vektor)}

\item{x.axis}{Beliebige X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}

\item{roots.bound.l}{Unterer Rand des Bereichs der Nullstellensuche}

\item{roots.bound.u}{Oberer Rand des Bereichs der Nullstellensuche}
}
\value{
cheb.list Liste berechneter Parameter (Koeffizienten, gefiltertes Modell, erste und zweite Ableitung des gefilterten Modells, Extremstellen und -Werte)
}
\description{
\code{cheb.fit.extr} fittet ein Chebyshev-Polynom beliebiger Ordnung an einen Datensatz/Zeitreihe mittels Least Squares Verfahren
}
