% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.fit}
\alias{cheb.fit}
\title{Curve Fitting with Chebyshev Polynomials}
\usage{
cheb.fit(d, x.axis, n)
}
\arguments{
\item{d}{Zu fittender Datensatz/Zeitreihe (Vektor)}

\item{x.axis}{Beliebige X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}
}
\value{
cheb.model Werte des gefilterten Modells
}
\description{
\code{cheb.fit} fittet ein Chebyshev-Polynom beliebiger Ordnung an einen Datensatz/Zeitreihe mittels Least Squares Verfahren
}
