% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.fit.seq}
\alias{cheb.fit.seq}
\title{Curve Fitting with Chebyshev Polynomials over Sequences}
\usage{
cheb.fit.seq(d, x.axis, n, l, bc.harmonic = FALSE)
}
\arguments{
\item{d}{Datensatz oder Zeitreihe}

\item{x.axis}{Beliebige X-Achse}

\item{n}{Ordnung des Polynoms}

\item{l}{Länge der Sequenzen}

\item{bc.harmonic}{Parameter für harmonische Randbedingung (Wahr / Falsch)}
}
\value{
cheb.model Modellwerte aus Least-Squares-Verfahren
}
\description{
Fittet ein Chebyshev Polynom beliebiger Ordnung an einen sequenzierten Datensatz/Zeitreihe mittels Least Squares Verfahren
}
