% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.rescale}
\alias{cheb.rescale}
\title{Rescaling of X-Axis (as vector or as scalar)}
\usage{
cheb.rescale(x.cheb, x.axis)
}
\arguments{
\item{x.cheb}{skalierte X-Achse (Skalar oder Vektor)}

\item{x.axis}{beliebige X-Achse (Vektor)}
}
\value{
x.rescaled reskalierte X-Achse (Skalar oder Sektor)
}
\description{
cheb.rescale reskaliert die für den Fit erzeugte Achse auf die Ursprüngliche
}
