% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.scale}
\alias{cheb.scale}
\title{Scaling of X-Axis (as vector or as scalar) to [-1,1]}
\usage{
cheb.scale(x.axis, x.val = NA)
}
\arguments{
\item{x.axis}{urspr체ngliche beliebige X-Achse (Vektor)}

\item{x.val}{Einzelwert auf X-Achse (Skalar)}
}
\value{
x.cheb.scaled Abh채ngig von Eingangsparametern: Skalierte X-Achse (Vektor) oder skalierter Einzelwert anhand vorgegebener X-Achse
}
\description{
\code{cheb.scale} skaliert beliebige X-Achse auf Achse, die f체r Polynom-fits vertr채glich ist.
}
