% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_1st}
\alias{cheb_1st}
\title{Generating Chebyshev Polynomials of first kind}
\usage{
cheb_1st(x_axis, n)
}
\arguments{
\item{x_axis}{beliebige X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}
}
\value{
cheb_t Chebyshev-Polynome Erster Art (Vektor)
}
\description{
cheb_1st erzeugt Chebyshev Polynome erster Art aus beliebiger X-Achse
}
