% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_2nd}
\alias{cheb_2nd}
\title{Generating Chebyshev Polynomials of second kind}
\usage{
cheb_2nd(x_axis, n)
}
\arguments{
\item{x_axis}{beliebigie X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}
}
\value{
cheb_u Chebyshev-Polynome Zweiter Art (Vektor)
}
\description{
cheb_2nd erzeugt Chebyshev Polynome zweiter Art aus beliebiger X-Achse
}
