% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_deriv_1st}
\alias{cheb_deriv_1st}
\title{Calculation of the values of the first derivation}
\usage{
cheb_deriv_1st(x_axis, cheb_coeff)
}
\arguments{
\item{x_axis}{beliebige X-Achse (Skalar oder Vektor)}

\item{cheb_coeff}{Chebyshev-Koeffizienten aus Least-Squares-Verfahren (Vektor)}
}
\value{
cheb_model_deriv_1st Erste Ableitung des gefilterten Modells (Skalar oder Vektor)
}
\description{
cheb_deriv_1st berechnet aus den Chebyshev-Koeffizienten die Werte der ersten Ableitung
}
