% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_find_extr}
\alias{cheb_find_extr}
\title{Curve Fitting with Chebyshev Polynomials and Finding of its Roots}
\usage{
cheb_find_extr(d, x_axis, n, roots_bound_lower = NA, roots_bound_upper = NA)
}
\arguments{
\item{d}{Zu fittender Datensatz/Zeitreihe (Vektor)}

\item{x_axis}{Beliebige X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}

\item{roots_bound_lower}{Unterer Rand des Bereichs der Nullstellensuche}

\item{roots_bound_upper}{Oberer Rand des Bereichs der Nullstellensuche}
}
\value{
cheb_list Liste berechneter Parameter (Koeffizienten, gefiltertes Modell, erste und zweite Ableitung des gefilterten Modells, Extremstellen und -Werte)
}
\description{
\code{cheb_fit.extr} fittet ein Chebyshev-Polynom beliebiger Ordnung an einen Datensatz/Zeitreihe mittels Least Squares Verfahren
}
