% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_find_max}
\alias{cheb_find_max}
\title{Routine to find maximum values of a curve}
\usage{
cheb_find_max(d, x_axis, n, maxima_bound_lower = NA,
  maxima_bound_upper = NA)
}
\arguments{
\item{d}{Zu fittender Datensatz}

\item{x_axis}{beliebiege X-Achse}

\item{n}{Ordnung des Skalars}

\item{maxima_bound_lower}{Unterer Rand des Bereichs der Maxima-Suche}

\item{maxima_bound_upper}{Oberer Rand des Bereichs der Maxima-Suche}
}
\value{
max_list Positionen und Werte der Maxima als Liste
}
\description{
\code{cheb_find_max} fittet ein Polynom und sucht mittel dessen Ableitung die Maxima des Datensatzes
}
