% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_fit}
\alias{cheb_fit}
\title{Curve Fitting with Chebyshev Polynomials}
\usage{
cheb_fit(d, x_axis, n)
}
\arguments{
\item{d}{Zu fittender Datensatz/Zeitreihe (Vektor)}

\item{x_axis}{Beliebige X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}
}
\value{
cheb_model Werte des gefilterten Modells
}
\description{
\code{cheb_fit} fittet ein Chebyshev-Polynom beliebiger Ordnung an einen Datensatz/Zeitreihe mittels Least Squares Verfahren
}
