% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_model_filter}
\alias{cheb_model_filter}
\title{Calculation of Values of the model fit}
\usage{
cheb_model_filter(x_axis, cheb_coeff)
}
\arguments{
\item{x_axis}{beliebige X-Achse (Skalar oder Vektor)}

\item{cheb_coeff}{Chebyshev-Koeffizienten aus Least-Squares-Verfahren (Vektor)}
}
\value{
cheb_model gefiltertes Modell (Skalar oder Vektor)
}
\description{
cheb_model berechnet aus den Chebyshev-Koeffizienten die Y-Werte
}
