% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_rescale}
\alias{cheb_rescale}
\title{Rescaling of X-Axis (as vector or as scalar)}
\usage{
cheb_rescale(x_cheb, x_axis)
}
\arguments{
\item{x_cheb}{skalierte X-Achse (Skalar oder Vektor)}

\item{x_axis}{beliebige X-Achse (Vektor)}
}
\value{
x_rescaled reskalierte X-Achse (Skalar oder Sektor)
}
\description{
cheb_rescale reskaliert die fuer den Fit erzeugte Achse auf die Urspruengliche
}
