% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.find.max}
\alias{cheb.find.max}
\title{Routine to find maximum values of a curve}
\usage{
cheb.find.max(d, x.axis, n, max.bound.l = NA, max.bound.u = NA)
}
\value{
max.list
}
\description{
\code{cheb.find.max} fittet ein Polynom und sucht mittel dessen Ableitung die Maxima des Datensatzes
}
